Finished the tutorial
It may need some editing
