--- app/Overview/css/overview.css
@@ -1,78 +1,0 @@
-/*	Style Guide:
- *		Transparent: 70% Opacity
- *
- * 		White:      #FFFFFF
- * 		Grey-White: #F5F5F5
- *		Faint-Grey: #ECECEC
- *		Light-Grey: #DDDDDD
- *		Grey:       #C5C5C5
- * 		Grey-Black: #4A4A4A
- * 		Black:      #000000
- */
-body {
-	margin: 0px;
-	padding: 0px;
-	border-spacing: 0px;
-	font-family: 'Roboto Condensed', sans-serif;
-	font-weight: 300;
-	font-size: 18px;
-}
-.header {
-	opacity: .7;
-	padding: 20px;
-	border-bottom: 2px solid #f5f5f5;
-	background-color: #4a4a4a;
-	color: #fff;
-	height: 50px;
-	vertical-align: top;
-}
-.header .title {
-	display: inline-block;
-	font-size: 32px;
-	color: #fff;
-}
-.header .capsule {
-	float: right;
-}
-.capsule {
-	display: inline-block;
-	border: 1px solid #ffffff;
-	border-radius: 4px;
-	overflow: hidden;
-}
-.capsule .pod {
-	display: inline-block;
-	padding: 10px;
-	padding-left: 16px;
-	padding-right: 16px;
-	border-right: 1px solid #ffffff;
-	color: #f5f5f5;
-}
-.capsule .pod:last-child {
-	border: none;
-	padding-right: 20px;
-}
-.frame {
-	color: #f5f5f5;
-	font-size: 32px;
-	padding-top: 18px;
-	padding-bottom: 0px;
-	padding-left: 30px;
-	padding-right: 16px;
-}
-.frame .text {
-	color: #c5c5c5;
-}
-.welcome {
-	margin-top: 50px;
-	text-align: center;
-}
-.welcome .large {
-	font-size: 72px;
-	color: #c5c5c5;
-}
-.welcome .small {
-	margin-top: 20px;
-	font-size: 32px;
-	color: #c5c5c5;
-}
